% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficientPlot.R
\name{coefficientPlot}
\alias{coefficientPlot}
\title{Server code for an RBrowse coefficient plot}
\usage{
coefficientPlot(input, output, session, data_file, overview)
}
\arguments{
\item{input}{The shiny input object. Passed automatically from \code{callModule}}

\item{output}{The shiny output object. Passed automatically from \code{callModule}}

\item{session}{The shiny session object. Passed automatically from \code{callModule}}

\item{data_file}{File path to a data file.}

\item{overview}{Returned data from calling the \code{\link{overviewPlot}} module. Needed to determine the
gene to plot}
}
\value{
passes through the output from \code{entityPlot}, ie. a reactive list containing
the plot
}
\description{
The \code{coefficientPlot} is a test run at configuring plots using the \code{\link{entityPlot}} module.
It will be used as an example for creating a number of modules designed to produce common
plots of data associated with a single genomic entity
It shows how to produce the necessary inputs for \code{entityPlot} and how to call it
}
