% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficientPlot.R
\name{coefficientPlotUI}
\alias{coefficientPlotUI}
\title{UI for an RBrowse coefficient plot}
\usage{
coefficientPlotUI(id)
}
\arguments{
\item{id}{Shiny id of the plot. This must be unique within your app.}
}
\value{
A set of shiny components comprising an RBrowse coefficient plot
}
\description{
The \code{coefficientPlot} is a test run at configuring plots using the \code{\link{entityPlot}} module
}
