% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityPlot.R
\name{entityPlotUI}
\alias{entityPlotUI}
\title{UI for an RBrowse entity plot}
\usage{
entityPlotUI(id)
}
\arguments{
\item{id}{Shiny id of the plot. This must be unique within your app.}
}
\value{
A set of shiny components comprising an RBrowse entity plot
}
\description{
The \code{entityPlot} provides a generic interface for producing plots from
genomic entities (genes, etc.).
}
