% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBrowse_demo.R
\name{RBrowse_demo}
\alias{RBrowse_demo}
\title{Run RBrowse demo}
\usage{
RBrowse_demo(demo = "simple-demo", ...)
}
\arguments{
\item{demo}{Name of the demo to run.}

\item{...}{Additional arguments passed to \code{\link[shiny]{runApp}}}
}
\description{
Run RBrowse demo
}
\examples{
if (interactive()) {
    RBrowse_demo()
}
}
