% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPlot.R
\name{testPlotUI}
\alias{testPlotUI}
\title{UI for an RBrowse test plot}
\usage{
testPlotUI(id)
}
\arguments{
\item{id}{Shiny id of the plot. This must be unique within your app.}
}
\value{
A set of shiny components comprising an RBrowse test plot
}
\description{
The \code{testPlot} is a test run at configuring plots using the \code{\link{trackedPlot}} module
}
