% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackedPlot.R
\name{trackedPlotUI}
\alias{trackedPlotUI}
\title{UI for an RBrowse tracked plot}
\usage{
trackedPlotUI(id)
}
\arguments{
\item{id}{}
}
\value{
A set of shiny components comprising an RBrowse tracked plot
}
\description{
The \code{trackedPlot} provides a generic interface for producing plots from
position based genomic data.
}
